Add module to implement Blackjack strategies
There are Blackjack strategy charts made of a 2D grid where a Player
can look up their Hand and the Dealer's card to find a recommended
strategy for the best outcome. These strategy charts are different for
possible split hands, soft hands with an ace, and every other hand. The
module applies the basic strategy as 2D-arrays of values that represent
a Blackjack action on a Player hand.