Update Decks class to be more memory and time efficient
Replaces two different arrays to hold cards used and unused with single array
to hold all cards and variable for moving index of current top card. This avoids
arrays allocating or deallocating memory frequently.
Using two different arrays to hold cards used and not used changes the length
of both arrays every time a card is taken from the deck. This can be slow since
memory is deallocated and allocated frequently in the arrays.
Create method that returns the fractional number betwen [0,1] representing the
number of cards left in the decks. This is a public method called outside the
Decks class to determine when to reshuffle.
Adds misc comments and docstrings.