Update Card React component
Adds boolean property to decide if a Card component can be flipped by a Player.
Changes default cardFlipState to 'Back' instead of either 'Front'/'Back'
depending on a React prop. The actually initial flipping occurs inside a
useEffect.
Adds timeoutId useRef to reference a setTimeout. This timeoutId useRef is
cleared inside a useEffect that runs when component unmounts.
Adds useRef to reference element of Card component that is translating from one
position to another in the DOM.
Uses a separate Card configuration when translating instead of the same
configuration for both static and translating.
Adds useEffect with dependency of cardFlipState that uses the cardElementRef
useRef to rotate the card element whenever fip state is changed.
Adds method to handle User clicking on card to flip it.
Fixes (deltaX, deltaY) coordinates returned by getDeckDeltaPosition to translate
from deck position to final card position.
Updates styling to fix issues with card not flipping when clicked.
Adds misc comments and docstrings.